move shared menu props into a common file
